* bugfix: warriors use 1000 rage as max instead of 100 (factor 10) * refactored statValues to their original dataTypes for consistency

+++ playerbot/TravelMgr.cpp
@@ -1003,7 +1003,7 @@ bool BossTravelDestination::isActive(Player* bot)
 
     int32 botLevel = bot->GetLevel();
 
-    uint8 botPowerLevel = AI_VALUE(uint8, "durability");
+	float botPowerLevel = AI_VALUE(float, "durability");
     float levelMod = botPowerLevel / 500.0f; //(0-0.2f)
     float levelBoost = botPowerLevel / 50.0f; //(0-2.0f)
