+++ playerbot/strategy/hunter/HunterActions.cpp
@@ -7,12 +7,12 @@ using namespace ai;
 
 bool CastSerpentStingAction::isUseful()
 {
-    return AI_VALUE2(uint8, "health", "current target") > 50;
+    return AI_VALUE2(float, "health", "current target") > 50;
 }
 
 bool CastViperStingAction::isUseful()
 {
-    return AI_VALUE2(uint8, "mana", "self target") < 50 && AI_VALUE2(uint8, "mana", "current target") >= 30;
+    return AI_VALUE2(float, "mana", "self target") < 50 && AI_VALUE2(float, "mana", "current target") >= 30;
 }
 
 bool CastAspectOfTheCheetahAction::isUseful()
