+++ playerbot/strategy/warrior/WarriorTriggers.cpp
@@ -7,6 +7,6 @@ using namespace ai;
 
 bool BloodrageBuffTrigger::IsActive()
 {
-    return AI_VALUE2(uint8, "health", "self target") >= sPlayerbotAIConfig.mediumHealth &&
-        AI_VALUE2(uint8, "rage", "self target") < 20;
+    return AI_VALUE2(float, "health", "self target") >= sPlayerbotAIConfig.mediumHealth &&
+        AI_VALUE2(uint32, "rage", "self target") < 20;
 }
